requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
bleach==6.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
cryptography==45.0.3
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-jazzmin==3.0.1
django-tinymce==4.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.4.0
filelock==3.18.0
fsspec==2025.5.1
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
langdetect==1.0.9
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pillow==10.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
tenacity==8.5.0
tiktoken==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
